Cleaning up some file organization and css
