+++ web/src/App.jsx
@@ -2,10 +2,10 @@ import { Routes, Route, Link } from "react-router-dom";
 import Container from "react-bootstrap/Container";
 import { NavMenu } from "./components/NavMenu";
 import { Footer } from "./components/Footer";
-import Detail from "./components/Detail";
-import Edit from "./components/Edit";
-import Landing from "./components/Landing";
-import Search from "./components/Search";
+import Detail from "./views/Detail";
+import Edit from "./views/Edit";
+import Landing from "./views/Landing";
+import Search from "./views/Search";
 import { useAuth0 } from "@auth0/auth0-react";
 import Loading from "./components/Loading";
 import history from "./utils/history";
