+++ web/src/components/SearchResults.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import Button from "./Button";
+import Button from "../components/Button";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
