+++ web/src/views/Edit.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { useLocation } from "react-router-dom";
-import Button from "./Button";
-import CountyDropdown from "./CountyDropdown";
-import CategoryDropdown from "./CategoryDropdown";
+import Button from "../components/Button";
+import CountyDropdown from "../components/CountyDropdown";
+import CategoryDropdown from "../components/CategoryDropdown";
 import Form from "react-bootstrap/Form";
 import Container from "react-bootstrap/esm/Container";
 import Col from "react-bootstrap/Col";
